KNO/kno_super_ai/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
websockets>=11.0.0
python-socketio[asyncio_client]>=5.9.0
pydantic>=1.10.0
aiofiles>=23.0.0
pyautogui>=0.9.55
Pillow>=9.0.0
paho-mqtt>=1.6.1
psutil>=5.9.0
requests>=2.28.0